# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope:
    properties:
      code:
        example: "0000"
        type: string
      data: {}
      message:
        example: ok
        type: string
    type: object
...
  github_com_scc749_nimbus-blog-api_internal_controller_http_v1_request.SendCode:
    properties:
      captcha:
        maxLength: 10
        minLength: 4
        type: string
      captcha_id:
        maxLength: 64
        minLength: 1
        type: string
      email:
        type: string
    required:
    - captcha
    - captcha_id
    - email
    type: object
...
  github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.AuthorInfo:
    properties:
      id:
        type: integer
      nickname:
        type: string
      specialization:
        type: string
    type: object
  github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.BaseCategory:
    properties:
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
  github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.BaseTag:
    properties:
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
...
  github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.CategoryDetail:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      post_count:
        type: integer
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.LikeInfo:
    properties:
      liked:
        type: boolean
      likes:
        type: integer
    type: object
...
  github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.PostDetail:
    properties:
      author:
        $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.AuthorInfo'
      author_id:
        type: integer
      category:
        $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.BaseCategory'
      content:
        type: string
      created_at:
        type: string
      excerpt:
        type: string
      featured_image:
        type: string
      id:
        type: integer
      is_featured:
        type: boolean
      like:
        $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.LikeInfo'
      meta_description:
        type: string
      meta_title:
        type: string
      published_at:
        type: string
      read_time:
        type: string
      slug:
        type: string
      status:
        type: string
      tags:
        items:
          $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.BaseTag'
        type: array
      title:
        type: string
      updated_at:
        type: string
      views:
        type: integer
    type: object
  github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.PostSummary:
    properties:
      author:
        $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.AuthorInfo'
      author_id:
        type: integer
      category:
        $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.BaseCategory'
      created_at:
        type: string
      excerpt:
        type: string
      featured_image:
        type: string
      id:
        type: integer
      is_featured:
        type: boolean
      like:
        $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.LikeInfo'
      published_at:
        type: string
      read_time:
        type: string
      slug:
        type: string
      status:
        type: string
      tags:
        items:
          $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.BaseTag'
        type: array
      title:
        type: string
      updated_at:
        type: string
      views:
        type: integer
    type: object
  github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.PostSummaryPage:
    properties:
      current_page:
        type: integer
      list:
        items:
          $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.PostSummary'
        type: array
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.TagDetail:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      post_count:
        type: integer
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: Nimbus Blog 后端 API（Admin + Public V1）。
  title: Nimbus Blog API
  version: "1.0"
paths:
...
  /v1/content/categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.CategoryDetail'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
      summary: 分类列表
      tags:
      - V1.Content
  /v1/content/posts:
    get:
      description: 可选携带 BearerAuth，以返回点赞状态（like.liked）。
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 分页大小
        in: query
        name: page_size
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 分类 ID 过滤
        in: query
        name: filter.category_id
        type: string
      - description: 标签 ID 过滤
        in: query
        name: filter.tag_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.PostSummaryPage'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
      summary: 文章列表
      tags:
      - V1.Content
...
  /v1/content/posts/{id}/likes:
    delete:
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.LikeInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
      security:
      - BearerAuth: []
      summary: 取消点赞文章
      tags:
      - V1.Content
    post:
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.LikeInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
      security:
      - BearerAuth: []
      summary: 点赞/取消点赞文章
      tags:
      - V1.Content
  /v1/content/posts/{slug}:
    get:
      description: 可选携带 BearerAuth，以返回点赞状态（like.liked）。
      parameters:
      - description: 文章 Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.PostDetail'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
      summary: 文章详情（按 slug）
      tags:
      - V1.Content
  /v1/content/tags:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_v1_response.TagDetail'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
      summary: 标签列表
      tags:
      - V1.Content
  /v1/email/send-code:
    post:
      consumes:
      - application/json
      parameters:
      - description: 邮箱与图形验证码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_v1_request.SendCode'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/github_com_scc749_nimbus-blog-api_internal_controller_http_shared.Envelope'
      summary: 发送邮箱验证码
      tags:
      - V1.Email
...
schemes:
- http
- https
securityDefinitions:
  AdminSession:
    in: cookie
    name: fiber_session
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"